requirements.txt
numpy
pyyaml
scikit-learn
matplotlib
prefetch_generator
wandb
torch
torchvision
hydra-core
hydra_colorlog 
omegaconf
tqdm
rich
autopep8
flake8
pre-commit
black
isort
pytest
scikit-image
monai
einops
opencv
timm
torchinfo